Refactor main loop into consensus engine structure
Move the main loop code into an engine module and start laying things
out like a consensus engine.
